Admin navbar designed

+++ src/app/api/db/linktree-links/route.js
@@ -1,8 +1,15 @@
 import connectMongoDb from "@/lib/mongodb";
 import {NextResponse} from "next/server";
 import Link from "@/models/link";
+import {getServerSession} from "next-auth";
+import {authOptions} from "@/app/api/auth/[...nextauth]/route";
 
 export async function POST(request)  { // POST metoda pro api přidává odkazy
+    const session = await getServerSession(authOptions);
+    if (!session) {
+        return NextResponse.json({message: "Unauthorized"},{status: 401});
+    }
+
     const {label, link} = await request.json()
     await connectMongoDb()
     await Link.create({label, link});
@@ -16,6 +23,11 @@ export async function GET(request)  {
 }
 
 export async function DELETE (request)  {
+    const session = await getServerSession(authOptions);
+    if (!session) {
+        return NextResponse.json({message: "Unauthorized"},{status: 401});
+    }
+
     const id = request.nextUrl.searchParams.get("id")
     await connectMongoDb()
     await Link.findByIdAndDelete(id)
